main.go: enable concurrent access to single device
This commit allows discovered devices to be allocated many times by
explicitly passing a `count` parameter. Whenever a count is not
supplied, count is assumed to be 1. This allows devices to be shared by
containers.
To accommodate this change, the --device flag has been changed to be a
JSON object to allow for easier parsing and extensibility.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>