e2e: add basic end-to-end test
This commit adds the first e2e test to the repo. This test is basic, it
creates a deployment that requires a Fuse device and ensures that the
workload can be scheduled. The test does not ensure that the correct
device was mounted or any other predicates related to scheduling,
however it is a foundation.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>