fix(CI): e2e tests
The recent changes to the build-system broke CI. The biggest issue is
that CI has been improperly configured all along, allowing commits that
break the plugin to merge because the e2e tests use `imagePullPolicy:
Always`, meaning that the image that is used in CI is not the artifact
built from the commit but the image in the container registry. This
allowed a commit to merge a change that broke the gRPC code.

This commit reverts the breaking change to the gRPC code and fixes CI so
that the e2e tests are forced to use an image build from the proposed
commit, namely `generic-device-plugin:test` and not one that lives in
the container registry.

Signed-off-by: squat <lserven@gmail.com>